Clean up bofire benchmarks; fix bugs
